Extract Chalice-independent code from index controller
