+++ .mypy.ini
@@ -57,7 +57,8 @@ modules =
     azul.digests,
     lambdas.indexer.app,
     azul.schemas,
-    azul.indexer.index_controller
+    azul.indexer.index_controller,
+    azul.indexer.index_queue_service
 packages =
     azul.openapi
